Rename sftpclient to remotehost
